breaks out symbolic and core pipeline functionality
1. Splits operations and arg parsing functions out into separate files
2. Splits pipeline tree functions and symbolic pipeline function out into separate files
